desacopla index js
